update template .envrc
Problem with the base usage of nix-direnv is that nix wants to copy over
every subdirectory of a flake.
(https://github.com/NixOS/nix/issues/5551)

Now this is fine on small 5mb projects, although a bit annoying, but
when you're dealing with LLVM and 20Gb of build artifacts you don't want
this.

So after looking at
https://github.com/nix-community/nix-direnv/issues/347 and tweaking
things a little to my own liking, this is the result